move ncwms to the wms directory, make it an extension of wmsEndpoint
