+++ src/index.ts
@@ -58,8 +58,8 @@ export type {
   StacEndpointInfo,
   StacItemsDocument,
 } from './stac/index.js';
-export { NcwmsEndpoint } from './ncwms/endpoint.js';
-export type { NcwmsLayerDetails, NcwmsMinMax } from './ncwms/endpoint.js';
+export { NcwmsEndpoint } from './wms/ncwms/endpoint.js';
+export type { NcwmsLayerDetails, NcwmsMinMax } from './wms/ncwms/endpoint.js';
 
 export {
   useCache,
