+++ src/wms/ncwms/model.ts
@@ -1,0 +1,13 @@
+export interface NcwmsLayerDetails {
+  scaleRange: [number, number];
+  palettes: string[];
+  defaultPalette?: string;
+  supportedStyles: string[];
+  units: string;
+  bbox: BoundingBox;
+}
+
+export interface NcwmsMinMax {
+  min: number;
+  max: number;
+}
